Rename RecordingManager and associated namespaces
