+++ application/Core/MORR/Session/Exceptions/AlreadyRecordingException.cs
@@ -1,4 +1,4 @@
-﻿namespace MORR.Core.Recording.Exceptions
+﻿namespace MORR.Core.Session.Exceptions
 {
     /// <summary>
     ///     Exception thrown if a recording is started while another recording is already running.
